initial port, broken code
