--- src/LarsonTech.Encryption.MsTest/appSettings.json
@@ -1,3 +1,0 @@
-{
-  "Thumbprint" :  "xxx" 
-}
