+++ src/LarsonTech.Encryption.Test/Usings.cs
@@ -1,0 +1,3 @@
+﻿global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using System;
+global using System.Text;
