+++ src/LarsonTech.Encryption.Test/appSettings.json
@@ -1,0 +1,3 @@
+{
+  "Thumbprint" :  "xxx" 
+}
