+++ src/LarsonTech.Encryption/LarsonTech - Backup.Encryption.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>netcoreapp1.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<Company>Larson</Company>
+		<Copyright>Daniel Larson</Copyright>
+	</PropertyGroup>
+	
+</Project>
