+++ src/LarsonTech.Encryption/LarsonTech.Encryption.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<Nullable>enable</Nullable>
 		<Company>Larson</Company>
 		<Copyright>Daniel Larson</Copyright>
