+++ src/LarsonTech.Encryption/RsaEncryptor.cs
@@ -1,35 +1,36 @@
-﻿using System;
-using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
-
-namespace Bct.Common.Encryption
+﻿namespace LarsonTech.Encryption;
+public class RsaEncryptor : IBinaryEncryptor, IDisposable
 {
-    public class RsaEncryptor : IBinaryEncryptor, IDisposable
+    private X509Certificate2 Certificate { get; }
+
+    public RsaEncryptor(X509Certificate2 cert)
     {
-        private X509Certificate2 Certificate { get; }
+        Certificate = cert ?? throw new ArgumentNullException(nameof(cert));
+    }
 
-        public RsaEncryptor(X509Certificate2 cert)
-        {
-            Certificate = cert ?? throw new ArgumentNullException();
-        }
+    public byte[] DecryptBytes(byte[] encryptedValue)
+    {
+        using var rsa = Certificate.GetRSAPrivateKey();
+        return rsa?.Decrypt(encryptedValue, RSAEncryptionPadding.OaepSHA512) ?? Array.Empty<byte>();
+    }
 
-        public byte[] DecryptBytes(byte[] encryptedValue)
-        {
-            using var RSA = Certificate.GetRSAPrivateKey();
-            return RSA.Decrypt(encryptedValue, RSAEncryptionPadding.Pkcs1);
-        }
+    public byte[] EncryptBytes(byte[] data)
+    {
+        using var rsa = Certificate.GetRSAPublicKey();
+        if (rsa == null)
+            throw new InvalidOperationException();
 
-        public byte[] EncryptBytes(byte[] data)
-        {
-            using var rsa = Certificate.GetRSAPublicKey();
-            if (rsa == null)
-                throw new InvalidOperationException();
-            return rsa.Encrypt(data, RSAEncryptionPadding.Pkcs1);
-        }
+        return rsa.Encrypt(data,  RSAEncryptionPadding.OaepSHA512);
+    }
 
-        public void Dispose()
-        {
-            ((IDisposable)Certificate).Dispose();
-        }
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        ((IDisposable)Certificate).Dispose();
     }
 }
