+++ src/LarsonTech.Encryption/SimpleAesEncryptor.cs
@@ -1,20 +1,19 @@
-﻿namespace Bct.Common.Encryption
+﻿namespace LarsonTech.Encryption;
+
+/// <summary>
+/// A simple AES encryptor for use when you know the private key or if you just want to test the library.
+/// </summary>
+public class SimpleAesEncryptor : AesEncryptor
 {
     /// <summary>
-    /// A simple AES encryptor for use when you know the private key or if you just want to test the library.
+    /// Creates an AES encryptor with a known key. If you need to generate the key pass in NULL. 
     /// </summary>
-    public class SimpleAesEncryptor : AesEncryptor
+    /// <param name="key">The binary key. You should NEVER EVER generate this manually.</param>
+    public SimpleAesEncryptor(byte[] key) : base(key)
     {
-        /// <summary>
-        /// Creates an AES encryptor with a known key. If you need to generate the key pass in NULL. 
-        /// </summary>
-        /// <param name="key">The binary key. You should NEVER EVER generate this manually.</param>
-        public SimpleAesEncryptor(byte[] key) : base(key)
-        {
-        }
+    }
 
-        public SimpleAesEncryptor() : base(CreateCryptograhicKey())
-        {
-        }
+    public SimpleAesEncryptor() : base(CreateCryptographicKey())
+    {
     }
 }
