+++ src/LarsonTech.Encryption/Usings.cs
@@ -1,0 +1,7 @@
+﻿global using System;
+global using System.Linq;
+global using System.IO;
+global using System.Text;
+global using System.Security.Cryptography;
+global using System.Diagnostics;
+global using System.Security.Cryptography.X509Certificates;
